# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: University Maintenance Request System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List service requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, assigned, in_progress, completed, rejected]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: submittedBy
          in: query
          schema:
            type: integer
        - name: assignedTo
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestList"

    post:
      operationId: createRequest
      tags: [requests]
      summary: Submit a new service request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestInput"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get service request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service request with assignment and logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequestDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Update a service request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUpdate"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"

    delete:
      operationId: deleteRequest
      tags: [requests]
      summary: Delete a service request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /requests/{id}/assign:
    post:
      operationId: assignRequest
      tags: [requests]
      summary: Assign a request to a maintenance officer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "200":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /requests/{id}/status:
    post:
      operationId: updateRequestStatus
      tags: [requests]
      summary: Update the status of a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdateInput"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusLog"
  /requests/{id}/logs:
    get:
      operationId: getRequestLogs
      tags: [requests]
      summary: Get status history for a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Status logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusLog"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    ServiceRequest:
      type: object
      required: [id, title, description, status, priority, submittedBy, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, assigned, in_progress, completed, rejected]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        location:
          type: ["string", "null"]
        evidenceUrl:
          type: ["string", "null"]
        submittedBy:
          type: integer
        submitterName:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        officerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ServiceRequestDetail:
      type: object
      required: [id, title, description, status, priority, submittedBy, createdAt, updatedAt, logs]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, assigned, in_progress, completed, rejected]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        location:
          type: ["string", "null"]
        evidenceUrl:
          type: ["string", "null"]
        submittedBy:
          type: integer
        submitterName:
          type: ["string", "null"]
        submitterEmail:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        officerName:
          type: ["string", "null"]
        assignment:
          oneOf:
            - $ref: "#/components/schemas/Assignment"
            - type: "null"
        logs:
          type: array
          items:
            $ref: "#/components/schemas/StatusLog"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RequestInput:
      type: object
      required: [title, description, priority]
      properties:
        title:
          type: string
          minLength: 5
        description:
          type: string
          minLength: 10
        categoryId:
          type: integer
        priority:
          type: string
          enum: [low, medium, high, urgent]
        location:
          type: string
        evidenceUrl:
          type: string
    RequestUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        priority:
          type: string
          enum: [low, medium, high, urgent]
        location:
          type: string
        evidenceUrl:
          type: string
    RequestList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRequest"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Assignment:
      type: object
      required: [id, requestId, officerId, assignedBy, assignedAt]
      properties:
        id:
          type: integer
        requestId:
          type: integer
        officerId:
          type: integer
        officerName:
          type: ["string", "null"]
        assignedBy:
          type: integer
        assignedByName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        assignedAt:
          type: string
          format: date-time
    AssignmentInput:
      type: object
      required: [officerId]
      properties:
        officerId:
          type: integer
        notes:
          type: string
    StatusLog:
      type: object
      required: [id, requestId, newStatus, changedBy, createdAt]
      properties:
        id:
          type: integer
        requestId:
          type: integer
        changedBy:
          type: integer
        changedByName:
          type: ["string", "null"]
        oldStatus:
          type: ["string", "null"]
        newStatus:
          type: string
          enum: [pending, assigned, in_progress, completed, rejected]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    StatusUpdateInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, assigned, in_progress, completed, rejected]
        note:
          type: string
...