# lib/api-zod/src/generated/types/serviceRequestDetail.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * University Maintenance Request System API
 * OpenAPI spec version: 0.1.0
 */
import type { Assignment } from './assignment';
import type { ServiceRequestDetailPriority } from './serviceRequestDetailPriority';
import type { ServiceRequestDetailStatus } from './serviceRequestDetailStatus';
import type { StatusLog } from './statusLog';

export interface ServiceRequestDetail {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  status: ServiceRequestDetailStatus;
  priority: ServiceRequestDetailPriority;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  evidenceUrl?: string | null;
  submittedBy: number;
  /** @nullable */
  submitterName?: string | null;
  /** @nullable */
  submitterEmail?: string | null;
  /** @nullable */
  assignedTo?: number | null;
  /** @nullable */
  officerName?: string | null;
  assignment?: Assignment | null;
  logs: StatusLog[];
  createdAt: Date;
  updatedAt: Date;
}