# lib/api-zod/src/generated/types/statusLog.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * University Maintenance Request System API
 * OpenAPI spec version: 0.1.0
 */
import type { StatusLogNewStatus } from './statusLogNewStatus';

export interface StatusLog {
  id: number;
  requestId: number;
  changedBy: number;
  /** @nullable */
  changedByName?: string | null;
  /** @nullable */
  oldStatus?: string | null;
  newStatus: StatusLogNewStatus;
  /** @nullable */
  note?: string | null;
  createdAt: Date;
}